[AIP-21] Move operators.slack_operator providers.slack.operators.slack
