+++ scripts/ci/pylint_todo.txt
@@ -160,7 +160,7 @@
 ./airflow/operators/python.py
 ./airflow/providers/amazon/aws/operators/s3_file_transform.py
 ./airflow/operators/s3_to_redshift_operator.py
-./airflow/operators/slack_operator.py
+./airflow/providers/slack/operators/slack.py
 ./airflow/providers/sqlite/operators/sqlite.py
 ./airflow/providers/amazon/aws/hooks/redshift.py
 ./airflow/providers/amazon/aws/operators/athena.py
